+++ src/main/java/com/aetherteam/aether/accessories/api/AccessoriesContainerOwner.java
@@ -1,5 +1,0 @@
-package com.aetherteam.aether.accessories.api;
-
-interface AccessoriesContainerOwner extends AccessoriesStorage {
-    void onContainerChanged(String slotName);
-}